Backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
ldap3==2.9.1
google-generativeai==0.8.4
email-validator==2.2.0
